# Retained rendering in Fuchsia PhysicalShapeLayer