@liyuqian 2018-11-13
@chinmaygarde @mikejurka : Please have a look at this PR again as I now changed EntityNode to be stored as unique_ptr according to https://fuchsia-review.googlesource.com/c/garnet/+/224197

See also FL-53 about the performance difference with or without this retained rendering. Here's a short summary:

Before retained rendering,
the RenderFrame in Scenic Compositor takes 22.9ms on average, the GPURasterizer::Draw in flutter takes 23.5ms on average.

After retained rendering,
the RenderFrame in Scenic Compositor takes 15.7ms on average, the GPURasterizer::Draw in flutter takes 7.9ms on average.