@liyuqian 2019-01-17
@chinmaygarde @mikejurka : please check the updated PR that (I think) correctly manages the scenic resources: (1) only SubmitSurface when the surface is no longer used in retained rendering; (2) let the SurfaceProducer own surfaces and entity nodes so they'll get destructed in the right thread.